Delta logic validation for Avaya ingestion

+++ Ingestion/Avaya/db_csv_validation_scala_day0.scala
@@ -1,50 +1,33 @@
+//Compare full extract CSV with DB = no mismatch
 //libraries
 import scala.util.control._
 import java.util.Properties
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.functions.{col,lower,when,length,concat,lit,trim}
-import org.apache.spark.sql.{DataFrame,Column}
-import org.apache.spark.sql.types.DataTypes
+import org.apache.spark.sql.functions.{col,lower,when,length,concat,lit,trim,unix_timestamp}
+import org.apache.spark.sql.{DataFrame,Column,Dataset,Row}
+import org.apache.spark.sql.types.{DataTypes,TimestampType}
 import scala.collection.immutable.List
 import java.time.LocalDateTime;
 
 val startTimeMillis = System.currentTimeMillis()
 
-//This function will find unmatched columns between source ahnd target dataframes. 
-//It takes source dataframe, target dataframe, source primary, target primary key
-//returns a list of Dataframes which has srcKeyColumnName, srcMismatchColumnName, tarKeyColumnName, tarMismatchColumnName
-def findUnmatchedColumns_5rows(src_df: DataFrame, tar_df: DataFrame, srcpk: String, tarpk: String): List[DataFrame] = {
-  var unmatchedDf: DataFrame = null
-  var unmatchedDf_temp: DataFrame = Seq.empty[(String, String, String, String, String)].toDF("primary_key", "source_column_name", "target_column_name", "source_value", "target_value")
-  val src_columns: List[String] = src_df.schema.fields.map(_.name).toList
-  val tar_columns: List[String] = tar_df.schema.fields.map(_.name).toList
-  
-  var mergedColumns: List[String] = src_columns zip tar_columns map { case (a, b) => a + "," + b }
-  val selectiveDifferences_df: List[DataFrame] = mergedColumns.map(mergedColumn => {// List[List[Dataset[Row]]]
-    var col_arry = mergedColumn.split(",")
-    val startTimeMillis_1: Long = System.currentTimeMillis()    
-    var src_rows:Dataset[Row] = src_df.select(srcpk, col_arry(0)).except(tar_df.select(tarpk, col_arry(1)))
-    var tar_rows:Dataset[Row] = tar_df.select(tarpk, col_arry(1)).except(src_df.select(srcpk, col_arry(0)))
-    val startTimeMillis_2: Long = System.currentTimeMillis()
-    var unmatchedCol_df:DataFrame = spark.emptyDataFrame
-    
-    if(src_rows.count > 0 || tar_rows.count > 0){
-      //This block is for getting columns which have unmatched data
-      var src_unmatchedrows_df: DataFrame = src_rows.toDF()
-      var tar_unmatchedrows_df: DataFrame = tar_rows.toDF()
-           
-      src_unmatchedrows_df.createOrReplaceTempView("src_unmatchedrows_df") 
-      tar_unmatchedrows_df.createOrReplaceTempView("tar_unmatchedrows_df") 
-      
-      unmatchedCol_df = spark.sql("select src_unmatchedrows_df.*,tar_unmatchedrows_df.* from src_unmatchedrows_df, tar_unmatchedrows_df where src_unmatchedrows_df." + srcpk + " = tar_unmatchedrows_df." + tarpk)
-      unmatchedCol_df.createOrReplaceTempView("unmatchedCol_df")       
-             
-    }
-    unmatchedCol_df    
-  })  
-    selectiveDifferences_df
+// Check whether column exists in dataframe
+def hasColumn(df: DataFrame, colName: String): Boolean = {
+  df.columns.contains(colName)
+}
+
+//This function will convert string timestamp in csv file from yyyy-MM-dd'T'HH:mm:ss.SSS'Z' to yyyy-MM-dd'T'HH:mm:ss.SSS+0000
+def csv_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss(df: DataFrame, colName: String): DataFrame = {  
+  var newDf = df.withColumn(colName, unix_timestamp(col(colName), "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'").cast(TimestampType))
+  return newDf
+}
+
+//This function will convert string timestamp in SQL database from yyyy-MM-dd'T'HH:mm:ss.SSS+0000 to yyyy-MM-dd'T'HH:mm:ss.SSS+0000
+def db_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss(df: DataFrame, colName: String): DataFrame = {
+  var newDf = df.withColumn(colName, unix_timestamp(col(colName), "yyyy-MM-dd'T'HH:mm:ss.SSS'+0000'").cast(TimestampType))
+  return newDf
 }
 
 //This function will replace empty string with null string in all columns in a dataframe
@@ -61,6 +44,18 @@ def replaceEmptyStringsWithNull_allColumns(df: DataFrame): DataFrame = {
       }
   )
 } 
+
+//This function will append zeroes to a column in a dataframe upto the columnLength
+//NewDataframe with modidied column will be returned
+def appendZeroes(df: DataFrame, columnName: String, columnLength: Int) : DataFrame = {
+  var newDf: DataFrame = null
+  var tempDf: DataFrame = df
+  for( counter <- 1 to columnLength){
+    tempDf = tempDf.withColumn(columnName,when(length(col(columnName))<counter,concat(lit("0"), col(columnName))).otherwise(col(columnName)))
+  } 
+  newDf = tempDf
+  return newDf
+}
 
 //This function will trim a single column in a dataframe
 def trimColumn(df: DataFrame, colName: String): DataFrame = {    
@@ -92,9 +87,17 @@ def dropColumn(df: DataFrame, colName: String): DataFrame = {
   return newDf
 }
 
+//This function will drop last column from a dataframe
+def dropLastColumn(df: DataFrame): DataFrame = {    
+  var newDf: DataFrame = null
+  newDf = df.drop(df.columns.last)  
+  return newDf
+}
+
 //This function will take dataframe that needs to be cleaned and a dataframe of known issues
 //It will clean the dataframe as per the known issues and return a cleaned dataframe
-def fixKnownDataIssues(tobecleaned_df: DataFrame, kwnissues_orgn_colnme_func_df: DataFrame): DataFrame = { 
+def fixKnownDataIssues(tobecleaned_df: DataFrame, kwnissues_orgn_colnme_func_df: DataFrame): DataFrame = {
+  //if(kwnissues_orgn_colnme_func_df.limit(1).collect().isEmpty)
   var column_list: List[String] = kwnissues_orgn_colnme_func_df.select("column_name").map(_.getString(0)).collect().toList
   var function_names = kwnissues_orgn_colnme_func_df.select("function").collect().map(_.getString(0)).toList
   var newDf: DataFrame = null
@@ -118,22 +121,77 @@ def fixKnownDataIssues(tobecleaned_df: DataFrame, kwnissues_orgn_colnme_func_df:
 def determineActionAndExecute(tobecleaned_df: DataFrame, column_name: String, func_name: String): DataFrame = { 
   var newDf: DataFrame = null
   
-  if(func_name.toLowerCase().trim().equals("drop_columns")){
-    newDf = dropColumn(tobecleaned_df, column_name)
-  }else if(func_name.toLowerCase().trim().equals("trim_columns")){
-    newDf = trimColumn(tobecleaned_df, column_name)
-  }else if(func_name.toLowerCase().trim().equals("replace_emptystring_with_null")){
-    newDf = replaceEmptyStringsWithNull_perColumn(tobecleaned_df, column_name)
+  if(hasColumn(tobecleaned_df, column_name)){
+    if(func_name.toLowerCase().trim().equals("drop_columns")){
+      newDf = dropColumn(tobecleaned_df, column_name)
+    }else if(func_name.toLowerCase().trim().equals("trim_columns")){
+      newDf = trimColumn(tobecleaned_df, column_name)
+    }else if(func_name.toLowerCase().trim().equals("replace_emptystring_with_null")){
+      newDf = replaceEmptyStringsWithNull_perColumn(tobecleaned_df, column_name)
+    }else if(func_name.toLowerCase().trim().equals("drop_last_column")){
+      newDf = dropLastColumn(tobecleaned_df)
+    }else if(func_name.toLowerCase().trim().contains("append_zeroes_upto_")){    
+      var lengthOfField:Int = (func_name.replaceAll("append_zeroes_upto_","")).toInt
+      newDf = appendZeroes(tobecleaned_df, column_name, lengthOfField)
+    }else if(func_name.toLowerCase().trim().equals("csv_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss")){
+      newDf = csv_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss(tobecleaned_df, column_name)
+    }else if(func_name.toLowerCase().trim().equals("db_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss")){
+      newDf = db_yyyy_mm_dd_t_hh_mm_ss_sssz_to_yy_mm_dd_t_hh_mm_ss_sss(tobecleaned_df, column_name)
+    }else{
+      newDf = tobecleaned_df
+      println("Skipping function. Invalid function name: " + func_name)      
+    }
+  }else{
+    newDf = tobecleaned_df
+    println("Column does not exist in Dataframe: " + column_name)
+    println("Skipping transformation: " + func_name)
   }  
   return newDf
+}
+
+//This function will find unmatched columns between source ahnd target dataframes. 
+//It takes source dataframe, target dataframe, source primary, target primary key
+//returns a list of Dataframes which has srcKeyColumnName, srcMismatchColumnName, tarKeyColumnName, tarMismatchColumnName
+def findUnmatchedColumns_5rows(src_df: DataFrame, tar_df: DataFrame, srcpk: String, tarpk: String): List[DataFrame] = {
+  var unmatchedDf: DataFrame = null
+  var unmatchedDf_temp: DataFrame = Seq.empty[(String, String, String, String, String)].toDF("primary_key", "source_column_name", "target_column_name", "source_value", "target_value")
+  val src_columns: List[String] = src_df.schema.fields.map(_.name).toList
+  val tar_columns: List[String] = tar_df.schema.fields.map(_.name).toList
+  
+  var mergedColumns: List[String] = src_columns zip tar_columns map { case (a, b) => a + "," + b }
+  val selectiveDifferences_df: List[DataFrame] = mergedColumns.map(mergedColumn => {// List[List[Dataset[Row]]]
+    var col_arry = mergedColumn.split(",")
+    val startTimeMillis_1: Long = System.currentTimeMillis()    
+    var src_rows:Dataset[Row] = src_df.select(srcpk, col_arry(0)).except(tar_df.select(tarpk, col_arry(1)))
+    var tar_rows:Dataset[Row] = tar_df.select(tarpk, col_arry(1)).except(src_df.select(srcpk, col_arry(0)))
+    val startTimeMillis_2: Long = System.currentTimeMillis()
+    var unmatchedCol_df:DataFrame = spark.emptyDataFrame
+    
+    if(src_rows.count > 0 || tar_rows.count > 0){
+      //This block is for getting columns which have unmatched data
+      var src_unmatchedrows_df: DataFrame = src_rows.toDF()
+      var tar_unmatchedrows_df: DataFrame = tar_rows.toDF()
+           
+      src_unmatchedrows_df.createOrReplaceTempView("src_unmatchedrows_df") 
+      tar_unmatchedrows_df.createOrReplaceTempView("tar_unmatchedrows_df") 
+      
+      unmatchedCol_df = spark.sql("select src_unmatchedrows_df.*,tar_unmatchedrows_df.* from src_unmatchedrows_df, tar_unmatchedrows_df where src_unmatchedrows_df." + srcpk + " = tar_unmatchedrows_df." + tarpk)
+      unmatchedCol_df.createOrReplaceTempView("unmatchedCol_df")       
+             
+    }
+    unmatchedCol_df.show(5,false)
+    unmatchedCol_df    
+  })  
+    selectiveDifferences_df
 }
 
  
 //DB Connection String
 val jdbcHostname = "10.216.2.181"
 val jdbcPort = 1433
-val jdbcDatabase = "CentralDWH_LS"  
-val jdbcUrl = s"jdbc:sqlserver://10.216.2.181:1433;database=CentralDWH_LS;"
+//val jdbcDatabase = "CentralDWH_LS"  
+//val jdbcUrl = s"jdbc:sqlserver://10.216.2.181:1433;database=CentralDWH_LS;"
+//val jdbcUrl = s"jdbc:sqlserver://10.216.2.181:1433;database=" + jdbcDatabase + ";"
 val connectionProperties = new Properties()
 val driverClass = "com.microsoft.sqlserver.jdbc.SQLServerDriver"
 
@@ -142,7 +200,7 @@ connectionProperties.put("user", s"CDLread")
 connectionProperties.put("password", dbutils.secrets.get(scope = "adbtstkv00", key = "avayadb"))    
 connectionProperties.setProperty("Driver", driverClass)      
 
-
+var isAutoAnalysisRequired: Boolean = false
 // CONNECTING TO THE LANDING ZONE
 dbutils.fs.refreshMounts()
 spark.conf.set("fs.azure.account.key.edplztstsa00.blob.core.windows.net", dbutils.secrets.get(scope = "adbtstkv00", key = "edplztstsa00key"))
@@ -151,11 +209,17 @@ val lz_blobName = "others"
 val sourcefolder = "avaya"
 
 // *****LOADING TABLE LIST*****
-var parenttablelistpath = "wasbs://test-parameters@edplztstsa00.blob.core.windows.net/table_list/db_csv/ingestion_avaya/DB_CSV_TestData.csv"
+var parenttablelistpath = "wasbs://test-parameters@edplztstsa00.blob.core.windows.net/table_list/db_csv/ingestion_avaya/DB_CSV_TestData_Day0_Delta.csv"
 var parentTable_list = spark.read.format("csv").option("header", "true").load(parenttablelistpath).toDF
 parentTable_list.createOrReplaceTempView("parentTable_list")
-var tableNames = parentTable_list.select("tablename").filter("sourcefolder='"+ sourcefolder +"'").collect().map(_(0)).toList
-var extractDates = parentTable_list.select("extract_date").filter("sourcefolder='"+ sourcefolder +"'").collect().map(_(0)).toList
+var tableNames:List[String] = parentTable_list.select("tablename").filter("sourcefolder='"+ sourcefolder +"'").collect().map(_.getString(0)).toList
+//var extractDates = parentTable_list.select("extract_date").filter("sourcefolder='"+ sourcefolder +"'").collect().map(_(0)).toList
+var day0_extractDates = parentTable_list.select("day0_extract_date").filter("sourcefolder='"+ sourcefolder +"'").collect().map(_(0)).toList
+var pr_lst_db_pks:List[String] = parentTable_list.select("db_pk_column").collect().map(_.getString(0)).toList
+var pr_lst_csv_pks:List[String] = parentTable_list.select("csv_pk_column").collect().map(_.getString(0)).toList
+var is_validation_reqd_lst:List[String] = parentTable_list.select("runValidation").collect().map(_.getString(0)).toList
+
+
 
 // *****LOADING KNOWN ISSUES LIST*****
 var knownissueslistpath = "wasbs://test-parameters@edplztstsa00.blob.core.windows.net/table_list/db_csv/ingestion_avaya/known_data_issues.csv"
@@ -165,11 +229,25 @@ knownissues_list.createOrReplaceTempView("knownissues_list")
 
 
 for (iterator <- 0 until tableNames.length){
+     var startTimeForTableMillis = System.currentTimeMillis()   
   
-     var srcTableName = tableNames(iterator)
-     var extractDate = extractDates(iterator)
-    
-     var startTimeForTableMillis = System.currentTimeMillis()
+     var is_validation_reqd = is_validation_reqd_lst(iterator)    
+     var srcTableName: String = tableNames(iterator)
+     var day0_extractDate = day0_extractDates(iterator)
+     if(is_validation_reqd.equals("yes")){       
+          
+     var jdbcDatabase = ""//"Users"//"CentralDWH_LS"
+     
+     if(srcTableName.equals("acs_results_withfcr_ls")){
+       jdbcDatabase = "Users"
+     }else{
+       jdbcDatabase = "CentralDWH_LS"
+     }
+     var jdbcUrl = s"jdbc:sqlserver://10.216.2.181:1433;database=" + jdbcDatabase + ";"
+  
+     
+     var extractDate = day0_extractDate     
+     
   
     // Read Avaya data into dataframe    
      var df_srcTable = spark.read.jdbc(jdbcUrl, "dbo." + srcTableName, connectionProperties)
@@ -177,6 +255,7 @@ for (iterator <- 0 until tableNames.length){
 
     //Read CSV into dataframe
      var blob_folderName = "wasbs://" + lz_blobName + "@" + strUsername + ".blob.core.windows.net/" + sourcefolder + "/" + srcTableName
+     
      var blob_folderName_extractPath = blob_folderName + "/1/extract_date=" + extractDate
      var csvFileNames = dbutils.fs.ls(blob_folderName_extractPath)     
             
@@ -246,18 +325,29 @@ for (iterator <- 0 until tableNames.length){
      
   
      println("Database table minus CSV comparison output using scala except function: ")
-     spark.sql("SELECT * FROM df_resultSrcTableMinCsv" ).show()
-     println("Database table minus CSV mismatch count : " + srcTable_minus_csv_count)
+     println("Database table minus CSV mismatch count : " + srcTable_minus_csv_count) 
+     spark.sql("SELECT * FROM df_resultSrcTableMinCsv" ).show(5,false)
+     
 
      println("CSV VS Database table comparison output using scala except function:")
-     spark.sql("SELECT * FROM df_resultCsvMinSrcTable" ).show()
      println("CSV minus Database table mismatch count : " + csv_minus_srcTable_count)    
+     spark.sql("SELECT * FROM df_resultCsvMinSrcTable" ).show(5,false)
+     
      
     if((srcTable_minus_csv_count != 0) || (csv_minus_srcTable_count != 0)){
       println("Database data not matched with CSV data - FAIL")
       println("Top 10 Sequence number not matching: ")
-      df_resultCsvMinSrcTable.select($"_c0").limit(10).show(5,false)
-      df_resultSrcTableMinCsv.select($"sid_key").limit(10).show(5,false)
+      var dbpk: String = pr_lst_db_pks(iterator)
+      var csvpk: String = pr_lst_csv_pks(iterator)
+      df_resultSrcTableMinCsv.select(dbpk).limit(10).show(5,false)
+      df_resultCsvMinSrcTable.select(csvpk).limit(10).show(5,false)
+      
+      if(isAutoAnalysisRequired){
+        println("Below is the dataframe of columns not matching: ")
+        var unmatched_cols_dfs:List[DataFrame] = findUnmatchedColumns_5rows(df_resultSrcTableMinCsv, df_resultCsvMinSrcTable, dbpk, csvpk)
+        unmatched_cols_dfs.filter(unmatched_cols_df => unmatched_cols_df.count() > 0).foreach(unmatched_cols_df => unmatched_cols_df.show)
+      }
+      
     }else{
       println("Database data matched with CSV data successfully - PASS")
     }     
@@ -265,6 +355,9 @@ for (iterator <- 0 until tableNames.length){
      var endTimeForTableMillis = System.currentTimeMillis()    
      println("Total time in minutes for current test case is: " + ((endTimeForTableMillis-startTimeForTableMillis)/(1000 * 60)).toFloat)
      println("\n******************END TESTCASE EXECUTION*************************\n")
+   }else{
+      println("Skipping validation. runValidation flag set to NO in testdata.csv for table: " + srcTableName)
+    }
 }
 val endTimeMillis = System.currentTimeMillis()
 println("Total time in minutes for entire test suite: " + ((endTimeMillis-startTimeMillis)/(1000 * 60)).toFloat)
